refactor endpoint network body helpers
